requirements.txt
Django==1.8.2
djangorestframework==3.1.3
drf-nested-fields==0.9.1
